init caching
Makes a folder called `Scratch_Cache` that stores Image_C2D data, so we don't have to keep the image data for every image in memory. Lots of problems currently, initial loading times are a lot longer, loading images take longer, and there's a storage concern with a bunch of cached images taking up storage in the SD card.